cache getUsers res + move to utils
